Update main.c - now allows multi-word messages
Now reads an entire line of input from the user for the message using fgets(), allowing for multi-word messages. The fgets() function is used to read a line of input, and message[strcspn(message, "\n")] = '\0' is used to remove the newline character from the end of the message.

By using fgets() to read an entire line, the main() function in main.c now supports messages consisting of more than one word.