Update main.c - Removed unnecessary code & prompt functionality update
Moved send and receive apple functions to a loop:
The code was updated to include a loop that sends the apple and receives the apple for each node in the ring.

Updated prompt messages:
The prompt messages in main.c were updated to be more descriptive and user-friendly. For example, "Enter the number of nodes in the ring" was changed to "Enter the number of nodes in the ring (k)".

Added prompt for destination node:
A prompt for the destination node was added to the loop to allow the user to specify the destination node for each message.